Removed former integration tests (which were written without TestContainers)
