+++ .github/workflows/build.yml
@@ -22,8 +22,8 @@ jobs:
         with:
           java-version: '8'
           distribution: 'adopt'
-      - name: Run Tests for specific issues
-        run: ./gradlew test --tests ch.admin.bar.siard2.cmd.issues.*
+      - name: Run unit tests
+        run: ./gradlew test
       - name: Run integration tests
         run: ./gradlew integrationTest
