+++ src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/ResourcesLoader.java
@@ -5,13 +5,6 @@ import java.net.URL;
 import java.util.Optional;
 
 public class ResourcesLoader {
-
-
-
-
-    public final static String ORACLE_INIT = "config/oracle/00_create_user.sql";
-
-
     private ResourcesLoader() {
     }
 
@@ -21,10 +14,7 @@ public class ResourcesLoader {
                 .getResource(resource));
 
         return urlToResource.map(url -> new File(url.getFile()))
+                .filter(File::exists)
                 .orElseThrow(() -> new IllegalArgumentException(String.format("Resource \"%s\" not found", resource)));
-    }
-
-    public static class SiardProjectExamples {
-
     }
 }
