+++ src/integrationTest/resources/config/mssql/issue-25.sql
@@ -1,0 +1,7 @@
+CREATE TYPE [dbo].[dtDay2] FROM [tinyint] NOT NULL;
+
+CREATE TABLE [dbo].[testtable]
+(
+    [ID] [int] IDENTITY (1,1) NOT NULL,
+    [DAY]     [dbo].[dtDay2]       NULL
+);
