--- src/test/java/ch/admin/bar/siard2/cmd/_MsSqlTestSuite.java
@@ -1,14 +1,0 @@
-package ch.admin.bar.siard2.cmd;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ 
-	MsSqlFromDbTester.class,
-	MsSqlToDbTester.class
-})
-public class _MsSqlTestSuite {
-
-}
