--- src/test/java/ch/admin/bar/siard2/cmd/_NwTestSuite.java
@@ -1,13 +1,0 @@
-package ch.admin.bar.siard2.cmd;
-
-import org.junit.runner.*;
-import org.junit.runners.*;
-import org.junit.runners.Suite.*;
-
-@RunWith(Suite.class)
-@SuiteClasses({ 
-	NorthwindFromDbTester.class
-})
-public class _NwTestSuite
-{
-}
