--- src/test/java/ch/admin/bar/siard2/cmd/_OeTestSuite.java
@@ -1,13 +1,0 @@
-package ch.admin.bar.siard2.cmd;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ 
-	OeFromDbTester.class
-})
-public class _OeTestSuite {
-
-}
