--- src/test/java/ch/admin/bar/siard2/cmd/_OracleTestSuite.java
@@ -1,14 +1,0 @@
-package ch.admin.bar.siard2.cmd;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ 
-	OracleFromDbTester.class,
-	OracleToDbTester.class
-})
-public class _OracleTestSuite {
-
-}
