--- src/test/java/ch/admin/bar/siard2/cmd/_PostgresTestSuite.java
@@ -1,14 +1,0 @@
-package ch.admin.bar.siard2.cmd;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ 
-	PostgresFromDbTester.class,
-	PostgresToDbTester.class
-})
-public class _PostgresTestSuite {
-
-}
